// src/_workspace/types/hr-user-leave/HrUserLeave.ts
export interface UserLeaveInterface {
  LEAVE_REQUEST_ID?: string | number
  STATUS?: string
  FLEX_TIME_REQUEST_DATE?: string
  LEAVE_ATTACHMENT?: boolean | string
  LEAVE_REQUEST_FILE_UPLOAD_ID?: number
  LEAVE_REQUEST_FILE_UPLOAD_NAME?: string
  LEAVE_REQUEST_FILE_UPLOAD_PATH?: string
  EMPLOYEE_CODE?: string
  EMPLOYEE_ID?: string
  EMPLOYEE_NAME?: string
  EMPLOYEE_SURNAME?: string
  SECT_NAME?: string
  EMPLOYEE_SECTION?: string
  LEAVE_TYPE?: string
  LEAVE_TYPE_ID?: number
  LEAVE_TYPE_DESCRIPTION_TH?: string
  LEAVE_TYPE_DESCRIPTION_EN?: string
  LEAVE_DATE?: string
  LEAVE_REQUEST_START_DATE?: string
  LEAVE_REQUEST_END_DATE?: string
  TIME?: string
  LEAVE_REQUEST_TIME?: string
  TOTAL_DAY_LEAVE?: number
  LEAVE_REQUEST_TOTAL_DAY?: number
  REASON?: string
  LEAVE_REQUEST_REASON?: string
  MODIFIED?: string
  MODIFIED_DATE?: string
  MODIFIED_BY?: string
  UPDATE_BY?: string
  CREATE_DATE?: string
  CREATE_BY?: string
  IS_APPROVED?: number
  INUSE?: number
}